modified file uploaded logic

+++ Server/src/controllers/product.controller.js
@@ -5,8 +5,9 @@ import { fileRemoveFromCloud,
         fileRemoveFromDisc,
         fileUploadInCloudinary
        } from "../utility/fileManage.js";
+import AppError from "../utility/customError.js";
 
-export const getAllProducts = async (req, res) => {
+export const getAllProducts = async (req, res, next) => {
     
      const { page, limit, starting,numberOfButtonPage } = req.query;
      
@@ -39,6 +40,7 @@ export const getAllProducts = async (req, res) => {
          .find({})
          .skip(skip)
          .limit(noOfItemPerPage);
+
        return res.status(200).json({
          success: true,
          messgae: "successfully fetched data",
@@ -48,12 +50,14 @@ export const getAllProducts = async (req, res) => {
        });
   
     } catch (error) {
-      console.log(error);
+
+      next(error);
+
     }
 
 };
 
-export const getUserProducts = async (req, res) => {
+export const getUserProducts = async (req, res, next) => {
     
   const userId = req.user._id;
   
@@ -65,23 +69,20 @@ export const getUserProducts = async (req, res) => {
       return res.status(200).json({
         success: true,
         message: "Products fetched successfully",
-        value : products
+        products,
       });
 
   } catch (error) {
    
-    console.log("user product not found ",error.message);
-    return res.status(500).json({
-      success: false,
-      message: "Error fetching products"
-      
-    });
+    next(error);
+  
   }
 };
 
-export const addProduct = async(req,res)=>{
+export const addProduct = async(req, res, next)=>{
 
     const id = req.user._id;
+    let product = null;
     
   // this will destructure the req.body object
   const { productName, description, category, price } = req.body;
@@ -90,32 +91,28 @@ export const addProduct = async(req,res)=>{
 
   //   this will check any field is empty or not if empty then simply send a error response
   if (!productName || !description || !category || !price ) {
-    return res.status(401).json({
-      success: false,
-      message: "please fill all field",
-    });
+
+    throw new AppError("please fill all field",401);
+    
   }
 
   // if user don't added any images
   if (!req.files.length) {
-    return res.status(401).json({
-      success: false,
-      message: "please add images ",
-    });
+
+    throw new AppError("please add images ", 401);
+    
   }
 
   //   if there is lest than three images has uploaded then simply return a error response
   if (req.files.length < 3) {
     
-    return res.status(401).json({
-      success: false,
-      message: "please select atleast 3 images",
-    });
+     throw new AppError("please select atleast 3 images", 401);
+   
   }
 
   const images = req.files;
 
-    const product = await productModel.create({
+     product = await productModel.create({
       productName,
       description,
       category,
@@ -124,8 +121,10 @@ export const addProduct = async(req,res)=>{
     });
 
     //   this function will upload all imgaes to cloudinary and simply set all the link to product collection named images array
-    const uploadImages = async () => {
-      for (const image of images) {
+    
+    
+    
+    const uploadImage = async (image) => {
         
         const productImage = {};
 
@@ -138,61 +137,80 @@ export const addProduct = async(req,res)=>{
        
         product.images.push(productImage);
 
-      }
     };
 
-    await uploadImages();
+    const uploadProductImages = async () => {
+
+      const allPromisses = images.map((image)=>{
+        return uploadImage(image);
+      })
+      
+       await Promise.all(allPromisses);
+    };
+
+    await uploadProductImages();
     await product.save();
 
     return res.status(200).json({
       success: true,
       message: "product added successfully",
-      value : product
+      product
     });
 
   }
    catch (error) {
 
-    console.log(error.message);
+    if(product){
 
-    await productModel.findByIdAndDelete(product._id);
+      await productModel.findByIdAndDelete(product._id);
 
-    return res.status(500).json({
-        success : false,
-        message : error.message
-    })
+    }
+
+    next(error);
 
   }
   finally{
 
-    
+    const deleteFileFromLocal = async()=>{
+       const fileRemovePromisses = req.files.map((image) => {
+         return fileRemoveFromDisc(image);
+       });
+       await Promise.all(fileRemovePromisses);
+    }
+
     if ( req.files.length){
-      fileRemoveFromDisc(req.files);
+
+      deleteFileFromLocal();
+     
     }
         
   }
 
 }
 
-export const updateProduct = async(req,res)=>{
+export const updateProduct = async(req, res, next)=>{
 
   const { productId } = req.params;
+  const  userId = req.user._id;
  
   let images = {};
 
   try {
 
-    const product = await productModel.findById(productId);
+    const product = await productModel.findOne({ _id: productId, userId });
+
+    if ( !product){
+      throw new AppError("no product found",400);
+    }
     
     // this will destructure the req.body object
     const { productName, description, category, price } = req.body;
    
       // this will check any field is empty or not if empty then simply send a error response
     if (!productName || !description || !category || !price) {
-      return res.status(401).json({
-        success: false,
-        message: "please fill all field",
-      });
+
+      throw new AppError("please fill all field", 401);
+
     }
 
     product.productName = productName;
@@ -203,110 +221,149 @@ export const updateProduct = async(req,res)=>{
     
 
     if (req.files.length) {
+
       //   if there is lest than three images has uploaded then simply return a error response
       if (req.files.length < 3) {
-        return res.status(401).json({
-          success: false,
-          message: "please select atleast 3 images",
-        });
+
+       throw new AppError("please select atleast 3 images", 401);
+
       }
-      const existedImagesInDb = product.images;
-      fileRemoveFromCloud(existedImagesInDb);
+
+      const productImages = product.images.map((imageObject) => {
+        return imageObject.public_id;
+      });
+
+      const removeproductImagesFromCloud = async (productsPubicIds) => {
+        const removePromisses = productsPubicIds.map((image_id) => {
+          return fileRemoveFromCloud(image_id);
+        });
+
+        await Promise.all(removePromisses);
+      };
+
+      await removeproductImagesFromCloud(productImages);
+
       product.images = [];
       images = req.files;
       
       //   this function will upload all imgaes to cloudinary and simply set all the link to product collection named images array
-      const uploadImages = async () => {
+      const uploadImage = async (image) => {
 
-        for (const image of images) {
-          
-          const productImage = {};
+        const productImage = {};
 
-          const result = await fileUploadInCloudinary(image);
-        
-          if (result) {
-            productImage.secure_url = result.secure_url;
-            productImage.public_id = result.public_id;
-          }
-        
-          product.images.push(productImage);
+        const result = await fileUploadInCloudinary(image);
 
+        if (result) {
+          productImage.secure_url = result.secure_url;
+          productImage.public_id = result.public_id;
         }
-      };
-    
 
-      console.log(product);
-      await uploadImages();
+        product.images.push(productImage);
+
+      };
+
+      const uploadProductImages = async () => {
+        const allPromisses = images.map((image) => {
+          return uploadImage(image);
+        });
+
+        await Promise.all(allPromisses);
+      };
+
+      await uploadProductImages();
       
     }
     
     await product.save();
+
     return res.status(200).json({
       success: true,
       message: "product updated successfully",
-      value : product
+      product
     });
   
  } catch (error) {
 
-  console.log(error);
-    return res.status(500).json({
-      success: false,
-      message: error.message,
-    });
+    next(error);
 
   } finally {
 
-    if(req.files.length){
-     fileRemoveFromDisc(req.files);
+    const deleteFileFromLocal = async () => {
+      const fileRemovePromisses = req.files.map((image) => {
+        return fileRemoveFromDisc(image);
+      });
+      await Promise.all(fileRemovePromisses);
+    };
+
+    if (req.files.length) {
+       deleteFileFromLocal();
     }
     
   }  
   
 }
 
-export const deleteProduct = async(req,res)=>{
-    const {productId} = req.params;
-    console.log(productId);
+export const deleteProduct = async(req, res, next)=>{
+
+    const { productId } = req.params;
+    
     try {
-        const result = await productModel.findByIdAndDelete(productId);
-        if (result) {
+
+        const product = await productModel.findByIdAndDelete(productId);
+
+        if (product) {
+
+          const productImages = product.images.map((imageObject) => {
+            return imageObject.public_id;
+          });
+
+          const removeproductImagesFromCloud = async (productsPubicIds) => {
+            const removePromisses = productsPubicIds.map((image_id) => {
+              return fileRemoveFromCloud(image_id);
+            });
+
+            await Promise.all(removePromisses);
+          };
+
+          await removeproductImagesFromCloud(productImages);
+
           return res.status(200).json({
             success: true,
             message: "product deleted successfully",
-            value : result
+            product,
           });
+        }else{
+
+          throw new AppError("no product found",400);
+
         }
              
         
     } catch (error) {
 
-        console.error(error.message);
-        return res.status(500).json({
-            success : false,
-            message : "product not deleted successfully..."
-        })
+        next(error);
     }
     
 }
 
-export const searchProducts = async (req,res)=>{
+export const searchProducts = async (req, res, next)=>{
+
   const { searchedProductName } = req.params;
-  console.log(searchedProductName);
   
   try {
     const products = await productModel.find({productName:{$regex:searchedProductName,$options:"i"}});
+    
     return res.status(200).json({
       success :true,
       products
     })
+
   } catch (error) {
-    console.log(error.message);
-    res.status(500).json({
-      success : false,
-      message : "not getting product ..."
-    })
+
+    next(error);
+
   }
+
 }
